Organize type info into dedicated module
