+++ crates/bevy_reflect/src/attributes.rs
@@ -240,7 +240,7 @@ pub(crate) use impl_custom_attribute_methods;
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::{enums::VariantInfo, type_info::Typed, TypeInfo};
+    use crate::{enums::VariantInfo, info::Typed, TypeInfo};
     use alloc::{format, string::String};
     use core::ops::RangeInclusive;
