+++ crates/bevy_reflect/src/impls/alloc/vec.rs
@@ -3,7 +3,7 @@ use bevy_reflect_derive::impl_type_path;
 use crate::impls::macros::impl_reflect_for_veclike;
 #[cfg(feature = "functions")]
 use crate::{
-    from_reflect::FromReflect, type_info::MaybeTyped, type_path::TypePath,
+    from_reflect::FromReflect, info::MaybeTyped, type_path::TypePath,
     type_registry::GetTypeRegistration,
 };
