+++ crates/bevy_reflect/src/impls/macros/set.rs
@@ -70,8 +70,8 @@ macro_rules! impl_reflect_for_hashset {
                 V: $crate::from_reflect::FromReflect + $crate::type_path::TypePath + $crate::type_registry::GetTypeRegistration + Eq + core::hash::Hash,
                 S: $crate::type_path::TypePath + core::hash::BuildHasher + Default + Send + Sync,
             {
-                fn get_represented_type_info(&self) -> Option<&'static $crate::type_info::TypeInfo> {
-                    Some(<Self as $crate::type_info::Typed>::type_info())
+                fn get_represented_type_info(&self) -> Option<&'static $crate::info::TypeInfo> {
+                    Some(<Self as $crate::info::Typed>::type_info())
                 }
 
                 #[inline]
@@ -141,15 +141,15 @@ macro_rules! impl_reflect_for_hashset {
                 }
             }
 
-            impl<V, S> $crate::type_info::Typed for $ty
+            impl<V, S> $crate::info::Typed for $ty
             where
                 V: $crate::from_reflect::FromReflect + $crate::type_path::TypePath + $crate::type_registry::GetTypeRegistration + Eq + core::hash::Hash,
                 S: $crate::type_path::TypePath + core::hash::BuildHasher + Default + Send + Sync,
             {
-                fn type_info() -> &'static $crate::type_info::TypeInfo {
+                fn type_info() -> &'static $crate::info::TypeInfo {
                     static CELL: $crate::utility::GenericTypeInfoCell = $crate::utility::GenericTypeInfoCell::new();
                     CELL.get_or_insert::<Self, _>(|| {
-                        $crate::type_info::TypeInfo::Set(
+                        $crate::info::TypeInfo::Set(
                             $crate::set::SetInfo::new::<Self, V>().with_generics($crate::generics::Generics::from_iter([
                                 $crate::generics::TypeParamInfo::new::<V>("V")
                             ]))
