+++ crates/bevy_reflect/src/set.rs
@@ -8,9 +8,9 @@ use bevy_platform::collections::{hash_table::OccupiedEntry as HashTableOccupiedE
 use bevy_reflect_derive::impl_type_path;
 
 use crate::{
-    generics::impl_generic_info_methods, hash_error, type_info::impl_type_methods, ApplyError,
-    Generics, PartialReflect, Reflect, ReflectKind, ReflectMut, ReflectOwned, ReflectRef, Type,
-    TypeInfo, TypePath,
+    generics::impl_generic_info_methods, hash_error, ty::impl_type_methods, ApplyError, Generics,
+    PartialReflect, Reflect, ReflectKind, ReflectMut, ReflectOwned, ReflectRef, Type, TypeInfo,
+    TypePath,
 };
 
 /// A trait used to power [set-like] operations via [reflection].
