+++ crates/bevy_reflect/src/tuple_struct.rs
@@ -7,7 +7,7 @@ use crate::generics::impl_generic_info_methods;
 use crate::{
     attributes::{impl_custom_attribute_methods, CustomAttributes},
     tuple::{DynamicTuple, Tuple},
-    type_info::impl_type_methods,
+    ty::impl_type_methods,
     ApplyError, Generics, PartialReflect, Reflect, ReflectKind, ReflectMut, ReflectOwned,
     ReflectRef, Type, TypeInfo, TypePath, UnnamedField,
 };
